fix barrier tests
